Replace use of r2d2 with redis-rs ConnectionManager
Using r2d2 with redis isn't really necessary, as the redis crate can
multiplex a connection, and in fact its the recommended approach for most
use cases: https://github.com/redis-rs/redis-rs/pull/388#issuecomment-911589177

Therefore this changes things to use redis-rs internally, with the intent
of adding an async API to this crate in the future. Additionally this
enables TLS for Redis connections, to support Redis servers that require
TLS.

This change is mostly API-compatible, but does change the underlying
RedisPool type (its still called "pool", but its actually a
ConnectionManager), and drops the exported type RedisPooledConnection.